Add createEmptyTempFilePath.py  - A seperate helper script
It fills up `relpaths` folder with files that needs to be filled manually. These files are going to be used to define some variables later.

Reason of creating such system is that said variables are defined in more than 1 batch file. Because their values are the same and those batch files can't interact using environment variables, a single place to contain the values born hence this commit.